Adds module for checking city support.
