+++ src/civilization/city_support/mod.rs
@@ -1,0 +1,1 @@
+pub mod plugin;
