+++ src/civilization/mod.rs
@@ -7,3 +7,4 @@ pub mod console;
 pub mod conflict;
 pub mod city_construction;
 pub mod remove_surplus;
+pub mod city_support;
