+++ src/civilization/remove_surplus/mod.rs
@@ -1,1 +1,2 @@
 pub mod plugin;
+mod systems;
